Improve EIP 1102 Support
My previous 1102 PR was not ideal: It was still using the
account-polling logic for the `enable()` function, which meant if a user
declined the login, it would keep re-requesting it in a loop.

This version replaces the account polling with a `Connect` button on the
no-accounts view, giving the user a graceful and patient place to
initiate the login process in a way that should be compatible with all
major web3 browsers.

This broke the tests again, sorry about that, I kindof just got this
good enough for my personal uses at the moment.